+++ wakame-mixin/src/main/kotlin/cc/mewcraft/wakame/bridge/codec/PaperCodecs.kt
@@ -1,4 +1,4 @@
-package cc.mewcraft.wakame.bridge.serialization.codec
+package cc.mewcraft.wakame.bridge.codec
 
 import com.mojang.serialization.Codec
 import net.minecraft.core.registries.BuiltInRegistries
