--- wakame-mixin/src/main/kotlin/cc/mewcraft/wakame/bridge/serialization/configurate/placeholder.kt
@@ -1,2 +1,0 @@
-package cc.mewcraft.wakame.bridge.serialization.configurate
-
